jyk/excel_decl_pipeline/requirements.txt
openai>=1.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
xlrd==1.2.0

